i/ml_task_extractor/requirements.txt
# Core NLP and ML libraries
spacy>=3.7.0
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Date and time processing
dateparser>=1.1.8
python-dateutil>=2.8.2

# Text processing and utilities
nltk>=3.8.1
textblob>=0.17.1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1

# Configuration and logging
pydantic>=2.4.0
loguru>=0.7.2
pyyaml>=6.0.1

# API and real-time processing
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=11.0.2
asyncio-mqtt>=0.13.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0

# Optional: For advanced features
sentence-transformers>=2.2.2
openai>=1.3.0